# config/documentation/swagger-admin.json
{
	"swagger": "2.0",
	"info": {
		"title": "HonorVet - Admin API",
		"description": "HonorVet - Admin API Documentation",
		"version": "1.0"
	},
	"produces": ["application/json"],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		"/admin/v1/users/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-user-detail-{id}-get",
				"tags": ["Users Management"],
				"description": "Get User details by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/recruiter/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-recruiter-detail-{id}-get",
				"tags": ["Recruiter Management"],
				"description": "Get User details by admin",

				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/paymentPlan/transaction": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-transaction-list-get",
				"tags": ["Admin Transaction"],
				"description": "Get Admin Transaction by recruiter",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search by PlanName and recruiterName",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "",
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "date",
						"value": ""
					},
					{
						"name": "planType",
						"in": "query",
						"description": "0 for subscription plans 1 for inventory plans",
						"type": "number",
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/paymentPlan/transaction/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-transaction-detail-{id}-get",
				"tags": ["Admin Transaction"],
				"description": "Get single Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/paymentPlan/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-create-POST",
				"tags": ["Payment Plan"],
				"description": "Add Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "planName",
						"in": "formData",
						"description": "Plan name is required",
						"type": "string",
						"value": ""
					},
					{
						"name": "planAmount",
						"in": "formData",
						"description": "Plan amount is required",
						"type": "",
						"value": ""
					},
					{
						"name": "planType",
						"in": "formData",
						"description": "0 for subscription plans 1 for inventory plans",
						"type": "number",
						"value": ""
					},
					{
						"name": "numberOfJob",
						"in": "formData",
						"description": "Number of jobs assigned for each plan is required",
						"type": "",
						"value": ""
					},
					{
						"name": "numberOfView",
						"in": "formData",
						"description": "Number of viwe assign in each plan is required",
						"type": "",
						"value": ""
					},
					{
						"name": "concurrentJobsAssignedEachPlan",
						"in": "formData",
						"description": "Number of concurrent jobs assigned in each plan is required",
						"type": "string",
						"value": ""
					},
					{
						"name": "nonConcurrentJobsAssignedEachPlan",
						"in": "formData",
						"description": "Number of non-concurrent jobs assigned in each plan is required",
						"type": "string",
						"value": ""
					},
					{
						"name": "validityType",
						"in": "formData",
						"description": "validityType",
						"type": "string",
						"value": ""
					},
					{
						"name": "description",
						"in": "formData",
						"description": "added in description",
						"type": "string",
						"value": ""
					},
					{
						"name": "validity",
						"in": "formData",
						"description": "Payment Plan validity is required",
						"type": "string",
						"value": ""
					},
					{
						"name": "duration",
						"in": "formData",
						"description": "Payment Plan validity is required",
						"type": "",
						"value": ""
					},
					{
						"name": "numberOfClicks",
						"in": "formData",
						"description": "Payment Plan numberOfClicks is required",
						"type": "",
						"value": ""
					},
					{
						"name": "jobType",
						"enum": [0, 1],
						"description": "0 for concurrent plans 1 for non-concurrent plans",
						"in": "formData",
						"type": "integer",
						"required": false
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-list-get",
				"tags": ["Payment Plan"],
				"description": "Get Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false,
						"value": "100"
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false,
						"value": "0"
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "Search on the bases of planName",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					},
					{
						"name": "startDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "",
						"value": ""
					},
					{
						"name": "endDate",
						"in": "query",
						"description": "YYYY-MM-DD",
						"required": false,
						"type": "date",
						"value": ""
					},
					{
						"name": "planType",
						"enum": [0, 1],
						"description": "0 for subscription plans 1 for inventory plans",
						"in": "query",
						"type": "integer",
						"required": false
					},
					{
						"name": "expireSoon",
						"value": "0",
						"description": "if you want only send ",
						"in": "query",
						"type": "number",
						"required": false
					},
					{
						"name": "dayType",
						"enum": ["DAY", "WEEK", "MONTH"],
						"description": "dayType ",
						"in": "query",
						"type": "string",
						"required": false
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-update-put",
				"tags": ["Payment Plan"],
				"description": "Update Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "planName",
						"in": "formData",
						"description": "Enter Plan name",
						"type": "string",
						"value": ""
					},
					{
						"name": "planAmount",
						"in": "formData",
						"description": "Enter the Plan amount.",
						"type": "",
						"value": ""
					},
					{
						"name": "numberOfJob",
						"in": "formData",
						"description": "Number of jobs assigned for each plan",
						"type": "",
						"value": ""
					},
					{
						"name": "numberOfView",
						"in": "formData",
						"description": "Number of viwe assign in each plan",
						"type": "",
						"value": ""
					},
					{
						"name": "concurrentJobsAssignedEachPlan",
						"in": "formData",
						"description": "Number of concurrent jobs assigned in each plan",
						"type": "string",
						"value": ""
					},
					{
						"name": "nonConcurrentJobsAssignedEachPlan",
						"in": "formData",
						"description": "Number of non-concurrent jobs assigned in each plan",
						"type": "string",
						"value": ""
					},
					{
						"name": "description",
						"in": "formData",
						"description": "added in description",
						"type": "string",
						"value": ""
					},
					{
						"name": "validity",
						"in": "formData",
						"description": "Payment Plan validity",
						"type": "string",
						"value": ""
					},
					{
						"name": "duration",
						"in": "formData",
						"description": "Payment Plan duration",
						"type": "",
						"value": ""
					},
					{
						"name": "numberOfClicks",
						"in": "formData",
						"description": "Payment Plan numberOfClicks",
						"type": "",
						"value": ""
					},
					{
						"name": "jobType",
						"enum": [0, 1],
						"description": "0 for concureent plans 1 for non-concuurent plans",
						"in": "formData",
						"type": "integer",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-delete-delete",
				"tags": ["Payment Plan"],
				"description": "Delete Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/paymentPlan/userDetail": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-paymentPlan-users-detail-list-get",
				"tags": ["Payment Plan UsersDetail"],
				"description": "Payment Plan UsersDetail",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "",
						"type": "string",
						"required": true
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "search",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/paymentPlan/block": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-block-update-put",
				"tags": ["Payment Plan"],
				"description": "Update Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/paymentPlan/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Payment-plan-detail-{id}-get",
				"tags": ["Payment Plan"],
				"description": "Get single Payment Plan by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		}
	},
...
}